refactor db folder
