+++ code/hyperdb/src/helpers/db/index.ts
@@ -1,0 +1,1 @@
+export * from './Db';
