--- code/hyperdb/src/helpers/db/main.ts
@@ -1,5 +1,0 @@
-import * as ipc from './main.ipc';
-
-export { ipc };
-export { Db } from './Db';
-export { create } from './main.create';
