+++ code/hyperdb/src/main/main.ts
@@ -1,26 +1,19 @@
-import * as db from '../helpers/db/main';
+import { Db } from '../helpers/db';
 import { Swarm } from '../helpers/swarm/main';
 import * as t from '../types';
+import * as ipc from './main.ipc';
 
-export { Swarm };
-export { Db } from '../helpers/db/main';
+export { Db, Swarm };
 export * from '../types';
 
 /**
- * Initializes a new HyperDB on the `main` process.
+ * Initializes a the `main` process for managing HyperDB requests from renderers.
  */
 export async function init(args: { ipc: t.IpcClient; log: t.ILog }) {
-  await db.ipc.init({ ipc: args.ipc, log: args.log });
+  await ipc.init({ ipc: args.ipc, log: args.log });
 }
 
 /**
- * Creates a new network connected HyperDB on the `main` process.
+ * Create a new network connected HyperDB on the `main` process.
  */
-export async function create(args: {
-  dir: string;
-  dbKey?: string;
-  autoAuth?: boolean;
-  join?: boolean;
-}) {
-  return db.create(args);
-}
+export { create } from './main.create';
